requirements.txt
streamlit
python-dotenv
openai
authlib
requests
gTTS
pandas
plotly